advance addding pending proposal number notification
